pyproject.toml
[project]
name = "game-of-life"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "HoopsMcann", email = ""}
]
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.5",
]

[project.scripts]
game-of-life = "game_of_life.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
